requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
blinker==1.8.2
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
constantly==23.10.4
cryptography==43.0.0
daphne==4.1.2
defusedxml==0.7.1
Django==4.2.15
dnspython==2.6.1
eventlet==0.36.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.4.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.4.0
locust==2.31.5
MarkupSafe==2.1.5
msgpack==1.0.8
oauthlib==3.2.2
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg==3.1.12
psycopg-binary==3.1.12
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pyzmq==26.2.0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.1
tomli==2.0.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.4
zipp==3.20.1
zope.event==5.0
zope.interface==7.0.1
